Moved the temporary dir change helper class into bud core.
